Made Home new component, fix links
