+++ src/App.js
@@ -1,11 +1,7 @@
 import React, { Component } from "react";
 import firebase from "./firebase";
 import Header from "./components/Header";
-import Tracker from "./components/Tracker";
-import Search from "./components/Search";
-import Results from "./components/Results";
-import Modal from "./components/Modal"
-import ReadingList from "./components/ReadingList";
+
 import "./styles/App.scss";
 import DisplayFirebase from "./components/DisplayFirebase";
 import AddComment from "./components/AddComment"
@@ -14,6 +10,7 @@ import {
   as Router, 
   Route, Link } 
   from 'react-router-dom';
+  import Home from './components/Home';
   
 
 class App extends Component {
@@ -27,19 +24,6 @@ class App extends Component {
       commentBookId: '',
       userGoal: {},
     };
-  }
-
-  componentDidMount() {
-    const dbRef = firebase.database().ref();
-    dbRef.on('value', (data) => {
-
-      //grab the data from FB, return an object
-      data = data.val();
-
-      //go through this object, and turn it into an array 
-      console.log(data)
-
-    })
   }
 
   //user input from search field
@@ -130,7 +114,15 @@ class App extends Component {
           <Link to="/bookshelf">Bookshelf</Link> */}
           <Route exact path='/' render={()=>{
             return(
-              <Home fullState={this.state}/>
+              <Home fullState={this.state} 
+                    appBookResults = {this.bookResults}
+                    appCloseModal = {this.closeModal}
+                    appOpenModal = {this.openModal}
+                    appSelectBook = {this.selectBook}
+                    appAddBook = {this.addBook}
+                    appComment = {this.handleComment}
+
+                    />
             )}
           } />
           <Route path="/bookshelf" component={DisplayFirebase} />
@@ -142,30 +134,3 @@ class App extends Component {
 
 export default App;
 
-class Home extends Component {
-  render(){
-    return(
-      <div>
-        <Tracker getGoalFn={this.goalFormSubmit}/>
-          <Search bookResults={this.bookResults} />
-          <Results displayBookResults={this.props.fullState.books}        
-                  selectBook={this.selectBook} />
-          {this.props.fullState.isShowing && (
-            <Modal
-              close={this.closeModal}
-              img={this.props.fullState.select.best_book.image_url}
-              title={this.props.fullState.select.best_book.title}
-              author={this.props.fullState.select.best_book.author.name}
-              rating={this.props.fullState.select.average_rating}
-              alt={this.props.fullState.select.best_book.title}
-              addBook={this.addBook}
-              selectBook = {this.props.fullState.select}
-            />
-          )}
-          <ReadingList addBook = {this.props.fullState.addBook} />
-          <Link to="/bookshelf">Bookshelf</Link>
-        {/* <AddComment comment={this.state.commentBookId} /> */}
-      </div>
-    )
-  }
-}
